Added newRelic/loadTime api endpoint
