+++ server/src/index.js
@@ -8,7 +8,7 @@ import chalk from 'chalk';
 
 import cache from './routes/cache';
 import bugsHistory from './routes/bugsHistory';
-import newRelicErrors from './routes/newRelicErrors';
+import newRelicRoute from './routes/newRelicRoute';
 import jiraIssues from './routes/jiraIssues';
 
 const app = express();
@@ -25,7 +25,7 @@ app.use(`${ROUTE_PREFIX}/cache`, cache);
 
 // widget endpoints
 app.use(`${ROUTE_PREFIX}/bugsHistory`, bugsHistory);
-app.use(`${ROUTE_PREFIX}/newRelicErrors`, newRelicErrors);
+app.use(`${ROUTE_PREFIX}/newRelic`, newRelicRoute);
 app.use(`${ROUTE_PREFIX}/jiraIssues`, jiraIssues);
 
 // catch 404 and forward to error handler
