+++ server/src/service/newRelicService.js
@@ -23,13 +23,22 @@ class NewRelicService {
     });
   }
 
-  getTransactionErrorsWeek() {
+  getTransactionErrors() {
     const nrql = 'SELECT count(*) FROM TransactionError WHERE appName = \'www.campsy.de\' SINCE 7 DAYS AGO COMPARE WITH 1 week ago';
 
     return this.getQueryResponse(nrql)
       .then((insightsResponse) => ({
         previous: _get(insightsResponse, 'previous.results[0].count'),
         current: _get(insightsResponse, 'current.results[0].count'),
+      }));
+  }
+
+  getLoadTime() {
+    const nrql = 'SELECT average(duration) from Transaction WHERE appName = \'www.campsy.de\' since 10 minutes ago';
+
+    return this.getQueryResponse(nrql)
+      .then((insightsResponse) => ({
+        current: _get(insightsResponse, 'results[0].average'),
       }));
   }
